refactor(advanced search): move select res. class comp.
