+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/resource-and-property-selection.component.ts
@@ -16,7 +16,7 @@ import {
     ResourceClassDefinition,
     ResourcePropertyDefinition
 } from '@dasch-swiss/dsp-js';
-import { SelectResourceClassComponent } from '../select-resource-class/select-resource-class.component';
+import { SelectResourceClassComponent } from './select-resource-class/select-resource-class.component';
 import { SelectPropertyComponent } from '../select-property/select-property.component';
 import { DspApiConnectionToken } from '../../../core/core.module';
