+++ projects/dsp-ui/src/lib/search/advanced-search/resource-and-property-selection/select-resource-class/select-resource-class.component.spec.ts
